🚚 rename @git-io -> socket-io.git
