+++ packages/socket-io.git/src/defaults.ts
